Build fix. Patch by Saleem Abdulrasool Fixes bug #568378
	* pkcs11/rpc-layer/gck-rpc-dispatch.c:
	* pkcs11/rpc-layer/gck-rpc-module.c: Build fix. Patch by  Saleem Abdulrasool
	Fixes bug #568378

svn path=/trunk/; revision=1488

+++ pkcs11/rpc-layer/gck-rpc-dispatch.c
@@ -202,7 +202,7 @@ proto_read_byte_array (CallState *cs, CK_BYTE_PTR* array, CK_ULONG* n_array)
 	GckRpcMessage *msg;
 	const unsigned char *data;
 	unsigned char valid;
-	uint32_t n_data;
+	size_t n_data;
 	
 	assert (cs);
